@@ -13,22 +13,22 @@
         }
     },
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^5.6 || ^7.0",
         "zendframework/zend-stdlib": "^3.0"
     },
     "require-dev": {
         "zendframework/zend-eventmanager": "^3.0",
         "zendframework/zend-inputfilter": "^2.6",
         "zendframework/zend-serializer": "^2.6.1",
-        "zendframework/zend-servicemanager": "^2.7.5 || ^3.0.3",
+        "zendframework/zend-servicemanager": "^3.0.3 || ^2.7.8",
         "zendframework/zend-filter": "^2.6",
-        "phpunit/phpunit": "^4.5",
+        "phpunit/phpunit": "^6.0.8 || ^5.7.15",
         "squizlabs/php_codesniffer": "^2.3.1"
     },
     "suggest": {
-        "zendframework/zend-eventmanager": "^2.6.2 || ^3.0, to support aggregate hydrator usage",
+        "zendframework/zend-eventmanager": "^3.0 || ^2.6.2, to support aggregate hydrator usage",
         "zendframework/zend-serializer": "^2.6.1, to use the SerializableStrategy",
-        "zendframework/zend-servicemanager": "^2.7.5 || ^3.0.3, to support hydrator plugin manager usage",
+        "zendframework/zend-servicemanager": "^3.0.3 || ^2.7.5, to support hydrator plugin manager usage",
         "zendframework/zend-filter": "^2.6, to support naming strategy hydrator usage"
     },
     "minimum-stability": "dev",
