@@ -9,7 +9,7 @@
 
 namespace ZendTest\Hydrator\Aggregate;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use stdClass;
 use Zend\EventManager\EventManager;
@@ -21,7 +21,7 @@ use Zend\Hydrator\HydratorInterface;
 /**
  * Unit tests for {@see AggregateHydrator}
  */
-class AggregateHydratorTest extends PHPUnit_Framework_TestCase
+class AggregateHydratorTest extends TestCase
 {
     /**
      * @var AggregateHydrator
