@@ -9,7 +9,7 @@
 
 namespace ZendTest\Hydrator\Aggregate;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use stdClass;
 use Zend\Hydrator\Aggregate\ExtractEvent;
 use Zend\Hydrator\Aggregate\HydrateEvent;
@@ -18,7 +18,7 @@ use Zend\Hydrator\Aggregate\HydratorListener;
 /**
  * Unit tests for {@see HydratorListener}
  */
-class HydratorListenerTest extends PHPUnit_Framework_TestCase
+class HydratorListenerTest extends TestCase
 {
     /**
      * @var \Zend\Hydrator\HydratorInterface|\PHPUnit_Framework_MockObject_MockObject
@@ -37,7 +37,7 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
      */
     public function setUp()
     {
-        $this->hydrator = $this->getMock('Zend\Hydrator\HydratorInterface');
+        $this->hydrator = $this->createMock('Zend\Hydrator\HydratorInterface');
         $this->listener = new HydratorListener($this->hydrator);
     }
 
@@ -46,7 +46,7 @@ class HydratorListenerTest extends PHPUnit_Framework_TestCase
      */
     public function testAttach()
     {
-        $eventManager = $this->getMock('Zend\EventManager\EventManagerInterface');
+        $eventManager = $this->createMock('Zend\EventManager\EventManagerInterface');
 
         $eventManager
             ->expects($this->exactly(2))
