@@ -10,7 +10,7 @@
 
 namespace ZendTest\Hydrator;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @requires PHP 5.4
