@@ -8,7 +8,7 @@
 namespace ZendTest\Hydrator;
 
 use Interop\Container\ContainerInterface;
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\HydratorInterface;
 use Zend\Hydrator\HydratorPluginManager;
 use Zend\Hydrator\HydratorPluginManagerFactory;
