@@ -9,6 +9,7 @@
 
 namespace ZendTest\Hydrator;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\ClassMethods;
 use Zend\Hydrator\Exception\BadMethodCallException;
 use Zend\Hydrator\Exception\InvalidArgumentException;
@@ -22,7 +23,7 @@ use ZendTest\Hydrator\TestAsset\ArraySerializable;
  *
  * @covers \Zend\Hydrator\ClassMethods
  */
-class ClassMethodsTest extends \PHPUnit_Framework_TestCase
+class ClassMethodsTest extends TestCase
 {
     /**
      * @var ClassMethods
@@ -81,7 +82,7 @@ class ClassMethodsTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetOptionsThrowsInvalidArgumentException()
     {
-        $this->setExpectedException(
+        $this->expectException(
             InvalidArgumentException::class,
             'The options parameter must be an array or a Traversable'
         );
@@ -106,7 +107,7 @@ class ClassMethodsTest extends \PHPUnit_Framework_TestCase
      */
     public function testExtractNonObjectThrowsBadMethodCallException()
     {
-        $this->setExpectedException(
+        $this->expectException(
             BadMethodCallException::class,
             'Zend\Hydrator\ClassMethods::extract expects the provided $object to be a PHP object)'
         );
@@ -118,7 +119,7 @@ class ClassMethodsTest extends \PHPUnit_Framework_TestCase
      */
     public function testHydrateNonObjectThrowsBadMethodCallException()
     {
-        $this->setExpectedException(
+        $this->expectException(
             BadMethodCallException::class,
             'Zend\Hydrator\ClassMethods::hydrate expects the provided $object to be a PHP object)'
         );
