@@ -9,10 +9,11 @@
 
 namespace ZendTest\Hydrator;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\HydratorInterface;
 use Zend\Hydrator\ClassMethods;
 
-class HydratorStrategyTest extends \PHPUnit_Framework_TestCase
+class HydratorStrategyTest extends TestCase
 {
     /**
      * The hydrator that is used during testing.
@@ -106,7 +107,7 @@ class HydratorStrategyTest extends \PHPUnit_Framework_TestCase
     ) {
         $hydrator = new ClassMethods($underscoreSeparatedKeys);
 
-        $strategy = $this->getMock('Zend\Hydrator\Strategy\StrategyInterface');
+        $strategy = $this->createMock('Zend\Hydrator\Strategy\StrategyInterface');
 
         $entity = new TestAsset\ClassMethodsUnderscore();
         $value = $entity->getFooBar();
