@@ -9,6 +9,7 @@
 
 namespace ZendTest\Hydrator\Strategy;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\Strategy\DateTimeFormatterStrategy;
 
 /**
@@ -16,7 +17,7 @@ use Zend\Hydrator\Strategy\DateTimeFormatterStrategy;
  *
  * @covers \Zend\Hydrator\Strategy\DateTimeFormatterStrategy
  */
-class DateTimeFormatterStrategyTest extends \PHPUnit_Framework_TestCase
+class DateTimeFormatterStrategyTest extends TestCase
 {
     public function testHydrate()
     {
@@ -58,7 +59,10 @@ class DateTimeFormatterStrategyTest extends \PHPUnit_Framework_TestCase
 
     public function testAcceptsStringCastableDateTimeFormat()
     {
-        $format = $this->getMock('stdClass', ['__toString']);
+        $format    = $this
+            ->getMockBuilder('stdClass')
+            ->setMethods(['__toString'])
+            ->getMock();
 
         $format->expects($this->once())->method('__toString')->will($this->returnValue('d/m/Y'));
 
