@@ -9,6 +9,7 @@
 
 namespace ZendTest\Hydrator\Strategy;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Hydrator\Strategy\ExplodeStrategy;
 
 /**
@@ -16,7 +17,7 @@ use Zend\Hydrator\Strategy\ExplodeStrategy;
  *
  * @covers \Zend\Hydrator\Strategy\ExplodeStrategy
  */
-class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
+class ExplodeStrategyTest extends TestCase
 {
     /**
      * @dataProvider getValidHydratedValues
@@ -40,7 +41,7 @@ class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
     {
         $strategy = new ExplodeStrategy();
 
-        $this->setExpectedException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
+        $this->expectException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
 
         $strategy->extract('');
     }
@@ -54,14 +55,14 @@ class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
 
     public function testGetExceptionWithEmptyDelimiter()
     {
-        $this->setExpectedException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
+        $this->expectException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
 
         new ExplodeStrategy('');
     }
 
     public function testGetExceptionWithInvalidDelimiter()
     {
-        $this->setExpectedException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
+        $this->expectException('Zend\Hydrator\Strategy\Exception\InvalidArgumentException');
 
         new ExplodeStrategy([]);
     }
@@ -79,7 +80,7 @@ class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
     {
         $strategy = new ExplodeStrategy();
 
-        $this->setExpectedException(
+        $this->expectException(
             'Zend\Hydrator\Strategy\Exception\InvalidArgumentException',
             'Zend\Hydrator\Strategy\ExplodeStrategy::hydrate expects argument 1 to be string,'
             . ' array provided instead'
@@ -92,7 +93,7 @@ class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
     {
         $strategy = new ExplodeStrategy();
 
-        $this->setExpectedException(
+        $this->expectException(
             'Zend\Hydrator\Strategy\Exception\InvalidArgumentException',
             'Zend\Hydrator\Strategy\ExplodeStrategy::hydrate expects argument 1 to be string,'
             . ' stdClass provided instead'
@@ -105,7 +106,7 @@ class ExplodeStrategyTest extends \PHPUnit_Framework_TestCase
     {
         $strategy = new ExplodeStrategy();
 
-        $this->setExpectedException(
+        $this->expectException(
             'Zend\Hydrator\Strategy\Exception\InvalidArgumentException',
             'Zend\Hydrator\Strategy\ExplodeStrategy::extract expects argument 1 to be array,'
             . ' stdClass provided instead'
